mascref/requirements.txt
Django==1.8.4
django-cors-headers==1.1.0
django-rest-auth==0.6.0
djangorestframework==3.3.0
django-rest-swagger==0.3.4
#mod-wsgi==4.4.21-ap24vc9
MySQL-python==1.2.3
PyYAML==3.11
six==1.10.0
django-crispy-forms==1.5.2
django-filter==0.11.0
drf-tracking==0.2.5
